+++ Source/Mosa.Compiler.Framework/CompilerStages/DebugFileGenerationStage.cs
@@ -3,11 +3,12 @@
 using Mosa.Compiler.Linker;
 using System.IO;
 
-namespace Mosa.Compiler.Framework.Stages
+namespace Mosa.Compiler.Framework.CompilerStages
 {
 	/// <summary>
 	/// An compilation stage, which generates a map file of the built binary file.
 	/// </summary>
+	/// <seealso cref="Mosa.Compiler.Framework.BaseCompilerStage" />
 	public sealed class DebugFileGenerationStage : BaseCompilerStage
 	{
 		#region Data members
