+++ Source/Mosa.Platform.x86/Intrinsic/FrameCallRetR8.cs
@@ -2,7 +2,7 @@
 
 using Mosa.Compiler.Framework;
 using Mosa.Compiler.Framework.IR;
-using Mosa.Platform.x86.MethodStages;
+using Mosa.Platform.x86.Stages;
 
 namespace Mosa.Platform.x86.Intrinsic
 {
