+++ Source/Mosa.Platform.x86/Stages/SimpleDeadCodeRemovalStage.cs
@@ -1,6 +1,6 @@
 // Copyright (c) MOSA Project. Licensed under the New BSD License.
 
-namespace Mosa.Platform.x86.MethodStages
+namespace Mosa.Platform.x86.Stages
 {
 	/// <summary>
 	/// The simple dead code removal stage remove useless instructions
